requirements.txt
absl-py==2.3.0
accelerate==0.31.0
addict==2.4.0
affine==2.4.0
aiofiles==24.1.0
aiohappyeyeballs==2.6.1
aiohttp==3.10.11
aiohttp-retry==2.8.3
aioice==0.10.1
aioresponses==0.7.8
aiortc==1.9.0
aiosignal==1.3.2
aiosqlite==0.21.0
altair==5.5.0
annotated-types==0.7.0
ansicon==1.89.0
anthropic==0.49.0
antlr4-python3-runtime==4.9.3
anyio==4.3.0
APScheduler==3.11.0
argon2-cffi==23.1.0
argon2-cffi-bindings==21.2.0
arrow==1.3.0
asgi-correlation-id==4.3.4
asgiref==3.7.2
asttokens==3.0.0
async-lru==2.0.4
asyncua==1.1.6
attrs==25.3.0
av==12.3.0
awscli==1.38.37
Babel==2.15.0
backoff==2.2.1
beautifulsoup4==4.12.3
bitarray==3.3.1
bitsandbytes==0.45.5
black==24.3.0
bleach==6.1.0
blessed==1.21.0
blinker==1.9.0
boto3==1.35.60
botocore==1.35.99
CacheControl==0.14.2
cachetools==5.5.0
cattrs==25.1.1
certifi==2024.2.2
cffi==1.16.0
charset-normalizer==3.3.2
click==8.1.7
click-plugins==1.1.1
cligj==0.7.2
codeflash==0.13.1
colorama==0.4.6
coloredlogs==15.0.1
comm==0.2.2
contourpy==1.2.1
coverage==7.8.2
crosshair-tool==0.0.92
cryptography==44.0.1
cycler==0.12.1
Cython==3.0.12
dataclasses-json==0.6.7
debugpy==1.8.14
decorator==5.2.1
defusedxml==0.7.1
dill==0.3.8
distlib==0.3.8
distro==1.9.0
Django==5.0.3
django-cors-headers==4.3.1
djangorestframework==3.15.0
dnspython==2.7.0
docker==7.1.0
docopt==0.6.2
docutils==0.19
editor==1.6.6
einops==0.8.0
elasticache-auto-discovery==1.0.0
executing==2.2.0
fastapi==0.110.3
fastapi-cprofile==0.0.2
fastjsonschema==2.19.1
filelock==3.13.1
filetype==1.2.0
firebase-admin==6.6.0
flake8==7.0.0
Flask==3.1.0
Flask-Cors==5.0.0
Flask-Mail==0.10.0
Flask-SQLAlchemy==3.1.1
flatbuffers==25.2.10
fonttools==4.53.1
fqdn==1.5.1
frozenlist==1.6.2
fsspec==2024.9.0
ftfy==6.3.1
git-filter-repo==2.47.0
gitdb==4.0.12
GitPython==3.1.44
google-ai-generativelanguage==0.6.10
google-api-core==2.21.0
google-api-python-client==2.149.0
google-auth==2.35.0
google-auth-httplib2==0.2.0
google-cloud-core==2.4.1
google-cloud-firestore==2.20.0
google-cloud-storage==3.0.0
google-crc32c==1.6.0
google-generativeai==0.8.3
google-resumable-media==2.7.2
googleapis-common-protos==1.65.0
greenlet==3.1.1
grpcio==1.66.2
grpcio-status==1.66.2
h11==0.14.0
httpcore==1.0.5
httplib2==0.22.0
httptools==0.6.4
httpx==0.28.1
huggingface-hub==0.32.4
humanfriendly==10.0
humanize==4.12.3
hydra-core==1.3.2
idna==3.7
ifaddr==0.2.0
imageio==2.37.0
imbalanced-learn==0.13.0
importlib_metadata==8.7.0
importlib_resources==6.5.2
inference==0.50.4
inference-development @ git+https://github.com/roboflow/inference.git@e1c3caf4619859b0bc523c6c6a9b2497adb8f0bb
iniconfig==2.1.0
inquirer==3.4.0
iopath==0.1.10
ipykernel==6.29.5
ipython==9.3.0
ipython_pygments_lexers==1.1.1
ipywidgets==8.1.2
isoduration==20.11.0
isort==5.13.2
itsdangerous==2.2.0
jax==0.6.1
jaxlib==0.6.1
jedi==0.19.2
Jinja2==3.1.4
jinxed==1.3.0
jiter==0.10.0
jmespath==1.0.1
joblib==1.4.2
json5==0.9.25
jsonpatch==1.33
jsonpointer==2.4
jsonschema==4.22.0
jsonschema-specifications==2023.12.1
junitparser==3.2.0
jupyter==1.0.0
jupyter-console==6.6.3
jupyter-events==0.10.0
jupyter-lsp==2.2.5
jupyter_client==8.6.3
jupyter_core==5.8.1
jupyter_server==2.14.0
jupyter_server_terminals==0.5.3
jupyterlab==4.1.8
jupyterlab_pygments==0.3.0
jupyterlab_server==2.27.1
jupyterlab_widgets==3.0.10
kiwisolver==1.4.5
langchain==0.3.25
langchain-core==0.3.64
langchain-text-splitters==0.3.8
langsmith==0.3.45
lazy_loader==0.4
libcst==1.8.0
line_profiler==4.2.0
lsprotocol==2023.0.1
lxml==5.4.0
markdown-it-py==3.0.0
MarkupSafe==2.1.5
marshmallow==3.26.1
matplotlib==3.9.1
matplotlib-inline==0.1.7
mccabe==0.7.0
mdurl==0.1.2
mediapipe==0.10.14
metlo==0.1.5
metlo-python-agent-bindings-common==0.0.2
mistune==3.0.2
ml_dtypes==0.5.1
mlxtend==0.23.1
mpmath==1.3.0
msgpack==1.1.0
multidict==6.4.4
mypy_extensions==1.1.0
mysql-connector==2.2.9
narwhals==1.26.0
nbclient==0.10.0
nbconvert==7.16.4
nbformat==5.10.4
nest-asyncio==1.6.0
networkx==3.4.1
notebook==7.1.3
notebook_shim==0.2.4
num2words==0.5.14
numpy==2.2.6
nvidia-ml-py==12.575.51
omegaconf==2.3.0
onnxruntime==1.21.1
openai==1.84.0
opencv-contrib-python==4.10.0.84
opencv-python==4.10.0.84
opencv-python-headless==4.10.0.84
opt_einsum==3.4.0
orjson==3.10.11
overrides==7.7.0
packaging==24.0
paho-mqtt==1.6.1
pandas==2.3.0
pandocfilters==1.5.1
parameterized==0.9.0
parso==0.8.4
pathspec==0.12.1
patsy==0.5.6
peft==0.11.1
pgmpy==0.1.26
pi==0.1.2
pillow==11.2.1
pillow_heif==0.22.0
pipenv==2023.12.1
platformdirs==4.3.8
plotly==6.0.0
pluggy==1.6.0
pmdarima==2.0.4
portalocker==3.1.1
posthog==4.4.0
prometheus-fastapi-instrumentator==6.0.0
prometheus_client==0.20.0
prompt_toolkit==3.0.51
propcache==0.3.1
proto-plus==1.24.0
protobuf==4.25.8
psutil==7.0.0
psycopg2-binary==2.9.10
pure_eval==0.2.3
py-cpuinfo==9.0.0
pyarrow==19.0.0
pyasn1==0.6.1
pyasn1_modules==0.4.1
pybase64==1.0.2
pycocotools==2.0.10
pycodestyle==2.11.1
pycparser==2.22
pydantic==2.9.2
pydantic-settings==2.7.1
pydantic_core==2.23.4
pydeck==0.9.1
pydot==2.0.0
pyee==13.0.0
pyflakes==3.2.0
pygls==1.3.1
Pygments==2.19.1
PyJWT==2.10.1
pylibsrtp==0.12.0
pylogix==1.0.5
pymemcache==4.0.0
pymodbus==3.8.3
pyOpenSSL==25.1.0
pyparsing==3.1.2
pyreadline3==3.5.4
pytest==8.4.0
pytest-asyncio==0.21.1
pytest-retry==1.6.3
pytest-timeout==2.4.0
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
python-json-logger==2.0.7
python-multipart==0.0.19
pytz==2023.3.post1
pyvips==3.0.0
pywin32==310
pywinpty==2.0.13
PyYAML==6.0.1
pyzmq==26.4.0
qtconsole==5.5.2
QtPy==2.4.1
rasterio==1.4.3
readchar==4.2.1
redis==5.0.8
referencing==0.35.1
regex==2024.11.6
requests==2.32.3
requests-file==2.1.0
requests-mock==1.11.0
requests-toolbelt==1.0.0
rf-clip==1.1
rf-groundingdino==0.2.0
rf-segment-anything==1.0
rfc3339-validator==0.1.4
rfc3986-validator==0.1.1
rich==13.9.4
roboflow==1.1.66
rpds-py==0.18.1
rpy2==3.5.17
rsa==4.7.2
runs==1.2.2
s3transfer==0.10.4
safetensors==0.5.3
samv2==0.0.4
scikit-image==0.25.2
scikit-learn==1.5.0
scipy==1.13.1
seaborn==0.13.2
Send2Trash==1.8.3
sentry-sdk==2.29.1
setuptools==80.9.0
shapely==2.0.7
shellingham==1.5.4
six==1.17.0
sklearn-compat==0.1.3
slack_sdk==3.33.5
smmap==5.0.2
sniffio==1.3.1
sortedcontainers==2.4.0
sounddevice==0.5.2
soupsieve==2.5
SQLAlchemy==2.0.36
sqlparse==0.4.4
stack-data==0.6.3
starlette==0.37.2
statsmodels==0.14.4
streamlit==1.42.0
structlog==24.4.0
supervision==0.25.1
sympy==1.13.1
tenacity==9.0.0
terminado==0.18.1
threadpoolctl==3.5.0
tifffile==2025.6.1
timeout-decorator==0.5.0
timm==1.0.15
tinycss2==1.3.0
tldextract==5.1.3
tokenizers==0.21.1
toml==0.10.2
tomlkit==0.13.3
torch==2.6.0
torchvision==0.21.0
tornado==6.5.1
tqdm==4.66.5
traitlets==5.14.3
transformers==4.51.3
twilio==9.3.8
typer==0.12.5
types-python-dateutil==2.9.0.20240316
typeshed_client==2.7.0
typing-inspect==0.9.0
typing_extensions==4.12.2
tzdata==2023.3
tzlocal==5.3.1
ultralytics==8.3.40
ultralytics-thop==2.0.14
unidiff==0.7.5
unittest-xml-reporting==3.2.0
uri-template==1.3.0
uritemplate==4.1.1
urllib3==2.2.1
uvicorn==0.22.0
virtualenv==20.25.1
watchdog==6.0.0
watchfiles==1.0.5
wcwidth==0.2.13
webcolors==1.13
webencodings==0.5.1
websocket-client==1.8.0
websockets==15.0.1
Werkzeug==3.1.3
wheel==0.45.1
widgetsnbextension==4.0.10
xgboost==2.1.1
xmod==1.8.1
yapf==0.43.0
yarl==1.20.0
z3-solver==4.15.0.0
zipp==3.23.0
zstandard==0.23.0
zxing-cpp==2.2.0
